+++ test/libsolidity/syntaxTests/events/event_struct_indexed.sol
@@ -1,0 +1,6 @@
+contract c {
+    struct S { uint a ; }
+    event E(S indexed);
+}
+// ----
+// TypeError: (51-52): This type is only supported in the new experimental ABI encoder. Use "pragma experimental ABIEncoderV2;" to enable the feature.