+++ Changelog.md
@@ -4,7 +4,8 @@ Important Bugfixes:
  * Code Generator: Properly perform cleanup for exponentiation and non-256 bit types.
  * Type Checker: Report error when using indexed structs in events with experimental ABIEncoderV2. This used to log wrong values.
  * Type Checker: Report error when using structs in events without experimental ABIEncoderV2. This used to crash or log the wrong values.
- * Parser: Treat unicode line endings as terminating strings and single-line comments.
+ * Parser: Consider all unicode line terminators (LF, VF, FF, CR, NEL, LS, PS) for single-line comments
+   and string literals. They are invalid in strings and will end comments.
  * Parser: Disallow unterminated multi-line comments at the end of input.
  * Parser: Treat ``/** /`` as unterminated multi-line comment.